+++ main.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/xmidt-org/chronon"
 	"go.uber.org/fx"
 )
 
@@ -11,7 +12,9 @@ func run(args []string) error {
 	app := fx.New(
 		parseCommandLine(args),
 		provideLogger(os.Stdout),
+		fx.Supply(chronon.SystemClock()),
 		provideActions(),
+		provideSwitchConfig(),
 		provideSwitch(),
 		provideHTTP(),
 	)